fix(studio): break all 7 circular dependency cycles and fix rules-of-hooks violation
