+++ packages/studio/src/utils/compositionPatterns.ts
@@ -1,0 +1,2 @@
+/** Matches the opening tag of a composition root element (e.g. `<div data-composition-id="main">`). */
+export const COMPOSITION_ROOT_OPEN_TAG_RE = /<[^>]*data-composition-id="[^"]+"[^>]*>/i;
